Extracting navigations components from containers
