+++ src/modules/category/containers/CategoriesScreen.js
@@ -1,19 +1,16 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { FlatList, View } from 'react-native';
+import { FlatList } from 'react-native';
 import { connect } from 'react-redux';
-
-import IonIcons from 'react-native-vector-icons/Ionicons';
 import { ItemList } from '../components/';
 import { Router } from '../../../router/routes';
-import { categoriesScreenStyles as styles } from '../../theme/categories/styles';
-import { colors } from '../../theme/styles';
 import { fetchCategories } from '../actions';
+import { rightSideIcon } from '../../navigator/components/NavigationBarItems';
 
 export class CategoriesScreen extends Component {
   constructor(props) {
     super(props);
-    this.onPressCategorie = this.onPressCategorie.bind(this);
+    this.onPressCategory = this.onPressCategory.bind(this);
     this.renderItem = this.renderItem.bind(this);
   }
 
@@ -21,15 +18,7 @@ export class CategoriesScreen extends Component {
     navigationBar: {
       visible: true,
       title: 'Etiquetas',
-      renderRight: () => {
-        return (
-          <View style={styles.navButtonsContainer}>
-            <View style={styles.navButton}>
-              <IonIcons name="ios-pricetag-outline" size={24} color={colors.whitePrimary} />
-            </View>
-          </View>
-        );
-      }
+      renderRight: rightSideIcon
     }
   }
 
@@ -38,22 +27,20 @@ export class CategoriesScreen extends Component {
   }
 
   renderItem({ item, index }) {
-    return <ItemList title={item.title} color={item.color} id={index} onPress={this.onPressCategorie} />;
+    return <ItemList title={item.title} color={item.color} id={index} onPress={this.onPressCategory} />;
   }
 
-  onPressCategorie(title, tab) {
+  onPressCategory(title, tab) {
     this.props.navigator.push(Router.getRoute('feed', { title, tab }));
   }
 
   render() {
     return (
-      <View>
-        <FlatList
-          renderItem={this.renderItem}
-          data={this.props.categories}
-          keyExtractor={(item, index) => index}
-        />
-      </View>
+      <FlatList
+        renderItem={this.renderItem}
+        data={this.props.categories}
+        keyExtractor={(item, index) => index}
+      />
     );
   }
 }
