+++ src/modules/feed/components/index.js
@@ -1,1 +1,6 @@
+export { CardBody } from './CardBody';
+export { CardFooter } from './CardFooter';
+export { CardHeader } from './CardHeader';
+export { CardSeparator } from './CardSeparator';
 export { FeedCard } from './FeedCard';
+export { FeedList } from './FeedList';
